refactor ui package
