+++ app/src/main/java/com/a7medelnoor/new_app/ui/first/FirstFragment.kt
@@ -1,4 +1,4 @@
-package com.a7medelnoor.new_app.ui.home
+package com.a7medelnoor.new_app.ui.first
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.a7medelnoor.new_app.R
 
-class SecondFragment : Fragment() {
+
+class FirstFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_second, container, false)
+        return inflater.inflate(R.layout.fragment_first, container, false)
     }
+
 }
