+++ app/src/main/res/layout/fragment_first.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.FirstFragment">
+    tools:context=".ui.first.FirstFragment">
 
     <TextView
         android:id="@+id/textView"
